+++ src/core/document.rs
@@ -333,7 +333,7 @@ fn print_directive_type_def(directive: &DirectiveDefinition) -> String {
     let locations = directive
         .locations
         .iter()
-        .map(|d| tailcall_typedefs_common::from_directive_location(d.node.clone()))
+        .map(|d| tailcall_typedefs_common::directive_definition::from_directive_location(d.node))
         .collect::<Vec<_>>();
     let repeatable = if directive.is_repeatable {
         " repeatable"
