+++ tailcall-typedefs-common/src/lib.rs
@@ -1,251 +1,19 @@
+use async_graphql_parser::types::{ServiceDocument, TypeSystemDefinition};
+
 mod common;
-mod field_type;
-
-use std::collections::{BTreeMap, HashSet};
-
-use async_graphql_parser::types::*;
-use async_graphql_parser::{Pos, Positioned};
-use async_graphql_value::Name;
-use common::get_description;
-use field_type::get_field_type;
-use schemars::schema::{RootSchema, Schema, SchemaObject};
-use schemars::JsonSchema;
-
-fn pos<A>(a: A) -> Positioned<A> {
-    Positioned::new(a, Pos::default())
-}
-
-#[derive(Clone)]
-pub struct Attrs {
-    pub name: &'static str,
-    pub repeatable: bool,
-    pub locations: Vec<&'static str>,
-}
-
-pub fn to_directive_location(str: &str) -> DirectiveLocation {
-    match str {
-        "Schema" => DirectiveLocation::Schema,
-        "Object" => DirectiveLocation::Object,
-        "FieldDefinition" => DirectiveLocation::FieldDefinition,
-        "EnumValue" => DirectiveLocation::EnumValue,
-        _ => DirectiveLocation::FieldDefinition,
-    }
-}
-
-pub fn from_directive_location(str: DirectiveLocation) -> String {
-    match str {
-        DirectiveLocation::Schema => String::from("SCHEMA"),
-        DirectiveLocation::Object => String::from("OBJECT"),
-        DirectiveLocation::FieldDefinition => String::from("FIELD_DEFINITION"),
-        DirectiveLocation::EnumValue => String::from("ENUM_VALUE_DEFINITION"),
-        _ => String::from("FIELD_DEFINITION"),
-    }
-}
-
-pub fn get_enum_values(obj: &Schema) -> Option<Vec<String>> {
-    match obj {
-        Schema::Object(schema_object) => {
-            if let Some(enum_values) = &schema_object.enum_values {
-                return Some(
-                    enum_values
-                        .iter()
-                        .map(|val| val.to_string())
-                        .collect::<Vec<String>>(),
-                );
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-fn generate_enum_definition(enum_values: Option<Vec<String>>, name: &str) -> TypeSystemDefinition {
-    let mut enum_values_defintions = vec![];
-    if let Some(enum_values) = enum_values {
-        for enum_value in enum_values {
-            let formated_value: String = enum_value
-                .to_string()
-                .chars()
-                .filter(|ch| ch != &'"')
-                .collect();
-            enum_values_defintions.push(pos(EnumValueDefinition {
-                value: pos(Name::new(formated_value)),
-                description: None,
-                directives: vec![],
-            }));
-        }
-    }
-
-    TypeSystemDefinition::Type(pos(TypeDefinition {
-        name: pos(Name::new(name)),
-        kind: TypeKind::Enum(EnumType { values: enum_values_defintions }),
-        description: None,
-        directives: vec![],
-        extend: false,
-    }))
-}
-
-fn generate_input_definition(schema: SchemaObject, name: &str) -> TypeSystemDefinition {
-    let description = get_description(&schema);
-
-    TypeSystemDefinition::Type(pos(TypeDefinition {
-        name: pos(Name::new(name)),
-        kind: TypeKind::InputObject(InputObjectType {
-            fields: generate_fields_definition(&schema),
-        }),
-        description: description.map(|inner| pos(inner.clone())),
-        directives: vec![],
-        extend: false,
-    }))
-}
-
-fn generate_fields_definition(schema: &SchemaObject) -> Vec<Positioned<InputValueDefinition>> {
-    let mut fields_definition = vec![];
-    if let Some(sub_schemas) = schema.subschemas.clone() {
-        if let Some(one_of) = sub_schemas.one_of {
-            for schema in one_of {
-                let schema_object = schema.into_object();
-                fields_definition.extend(do_stuff(&schema_object));
-            }
-
-            return fields_definition;
-        }
-    }
-
-    do_stuff(schema)
-}
-
-fn do_stuff(schema: &SchemaObject) -> Vec<Positioned<InputValueDefinition>> {
-    let mut arguments = vec![];
-    if let Some(properties) = schema
-        .object
-        .as_ref()
-        .map(|object| object.properties.clone())
-    {
-        for (name, property) in properties.into_iter() {
-            let property = property.into_object();
-            let description = get_description(&property);
-            let definition = pos(InputValueDefinition {
-                description: description.map(|inner| pos(inner.to_owned())),
-                name: pos(Name::new(name.to_owned())),
-                ty: pos(get_field_type(name, property.clone())),
-                default_value: None,
-                directives: Vec::new(),
-            });
-
-            arguments.push(definition);
-        }
-    }
-
-    arguments
-}
-
-pub trait DirectiveDefinition {
-    fn into_schemars() -> RootSchema
-    where
-        Self: JsonSchema,
-    {
-        schemars::schema_for!(Self)
-    }
-
-    fn directive_definition(generated_types: &mut HashSet<String>) -> Vec<TypeSystemDefinition>;
-
-    fn into_directive_definition(
-        root_schema: RootSchema,
-        attrs: Attrs,
-        generated_types: &mut HashSet<String>,
-    ) -> Vec<TypeSystemDefinition> {
-        let mut service_doc_definitions = vec![];
-        let definitions: BTreeMap<String, Schema> = root_schema.definitions;
-        let schema: SchemaObject = root_schema.schema;
-        let description = get_description(&schema);
-
-        for (name, schema) in definitions.iter() {
-            if generated_types.contains(name) {
-                continue;
-            }
-            let enum_values = get_enum_values(schema);
-            if enum_values.is_some() {
-                service_doc_definitions.push(generate_enum_definition(enum_values, name));
-                generated_types.insert(name.to_string());
-            } else {
-                service_doc_definitions.push(generate_input_definition(
-                    schema.clone().into_object(),
-                    name,
-                ));
-                generated_types.insert(name.to_string());
-            }
-        }
-
-        let directve_definition = TypeSystemDefinition::Directive(pos(
-            async_graphql_parser::types::DirectiveDefinition {
-                description: description.map(|inner| pos(inner.clone())),
-                name: pos(Name::new(attrs.name)),
-                arguments: generate_fields_definition(&schema),
-                is_repeatable: attrs.repeatable,
-                locations: attrs
-                    .locations
-                    .into_iter()
-                    .map(|val| pos(to_directive_location(val)))
-                    .collect(),
-            },
-        ));
-        service_doc_definitions.push(directve_definition);
-        service_doc_definitions
-    }
-}
-
-pub trait ScalarDefinition {
-    fn into_schemars() -> RootSchema
-    where
-        Self: JsonSchema,
-    {
-        schemars::schema_for!(Self)
-    }
-
-    fn scalar_definition() -> TypeSystemDefinition;
-
-    fn into_scalar_definition(root_schema: RootSchema, name: &str) -> TypeSystemDefinition {
-        let schema: SchemaObject = root_schema.schema;
-        let description = get_description(&schema);
-        TypeSystemDefinition::Type(pos(TypeDefinition {
-            name: pos(Name::new(name)),
-            kind: TypeKind::Scalar,
-            description: description.map(|inner| pos(inner.clone())),
-            directives: vec![],
-            extend: false,
-        }))
-    }
-}
-
-pub trait InputDefinition {
-    fn into_schemars() -> RootSchema
-    where
-        Self: JsonSchema,
-    {
-        schemars::schema_for!(Self)
-    }
-
-    fn input_definition() -> TypeSystemDefinition;
-
-    fn into_input_definition(root_schema: RootSchema, name: &str) -> TypeSystemDefinition {
-        let schema: SchemaObject = root_schema.schema;
-        let description = get_description(&schema);
-
-        TypeSystemDefinition::Type(pos(TypeDefinition {
-            name: pos(Name::new(name)),
-            kind: TypeKind::InputObject(InputObjectType {
-                fields: generate_fields_definition(&schema),
-            }),
-            description: description.map(|inner| pos(inner.clone())),
-            directives: vec![],
-            extend: false,
-        }))
-    }
-}
+pub mod directive_definition;
+mod enum_definition;
+pub mod input_definition;
+pub mod scalar_definition;
 
 pub struct ServiceDocumentBuilder {
     definitions: Vec<TypeSystemDefinition>,
+}
+
+impl Default for ServiceDocumentBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl ServiceDocumentBuilder {
@@ -278,7 +46,7 @@ impl ServiceDocumentBuilder {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use schemars::JsonSchema;
 
     #[derive(JsonSchema)]
     enum Schema {
